+++ packages/zowe-explorer-api/__tests__/__unit__/profiles/ZoweExplorerZosmfApi.unit.test.ts
@@ -89,6 +89,29 @@ describe("ZosmfUssApi", () => {
         jest.clearAllMocks();
     });
 
+    describe("_getSession", () => {
+        const exampleProfile = {
+            message: "",
+            type: "zosmf",
+            failNotFound: false,
+            name: "test.zosmf",
+            profile: {
+                host: "localhost",
+                password: "password",
+                protocol: "http",
+                user: "aZosmfUser",
+            },
+        } as zowe.imperative.IProfileLoaded;
+
+        it("should include profile properties in the built session object", () => {
+            const api = new ZoweExplorerZosmf.UssApi();
+
+            const transformedProps = { ...exampleProfile.profile, hostname: exampleProfile.profile?.host };
+            delete transformedProps["host"];
+            expect((api as any)._getSession(exampleProfile).mISession).toMatchObject(transformedProps);
+        });
+    });
+
     describe("updateAttributes", () => {
         const ussApi = new ZoweExplorerZosmf.UssApi();
         const getSessionMock = jest.spyOn(ussApi, "getSession").mockReturnValue(fakeSession);
@@ -262,16 +285,17 @@ describe("ZosmfUssApi", () => {
         expect(checkStatusSpy).toHaveBeenCalledTimes(1);
     });
 
-    it("should test that copy calls zowe.Utilities.putUSSPayload", () => {
+    it("should test that copy calls zowe.Utilities.putUSSPayload", async () => {
         const api = new ZoweExplorerZosmf.UssApi();
         api.getSession = jest.fn();
+        const response = Buffer.from("hello world!");
 
         Object.defineProperty(zosfiles.Utilities, "putUSSPayload", {
-            value: jest.fn().mockResolvedValue(Buffer.from("hello world!")),
+            value: jest.fn().mockResolvedValue(response),
             configurable: true,
         });
 
-        expect(api.copy("/")).toStrictEqual(zosfiles.Utilities.putUSSPayload(api.getSession(), "/", { request: "copy" }));
+        await expect(api.copy("/")).resolves.toEqual(response);
     });
 
     it("getStatus should validate unverified profile", async () => {