+++ packages/zowe-explorer/__tests__/__unit__/dataset/actions.unit.test.ts
@@ -2055,7 +2055,7 @@ describe("Dataset Actions Unit Tests - Function copyDataSets", () => {
             parentNode: blockMocks.datasetSessionNode,
         });
         child.contextValue = globals.DS_DS_CONTEXT;
-        await expect(dsActions.copyDataSets(child, null, blockMocks.testDatasetTree)).toStrictEqual(Promise.resolve());
+        await expect(dsActions.copyDataSets(child, null as any, blockMocks.testDatasetTree)).resolves.not.toThrow();
     });
     it("Checking copy of sequential datasets", async () => {
         globals.defineGlobals("");
@@ -2092,7 +2092,7 @@ describe("Dataset Actions Unit Tests - Function copyDataSets", () => {
             return Promise.resolve(prm);
         });
         expect(mocked(Gui.errorMessage)).not.toHaveBeenCalled();
-        await expect(dsActions.copyDataSets(nodeCopy, null, blockMocks.testDatasetTree)).toStrictEqual(Promise.resolve());
+        await expect(dsActions.copyDataSets(nodeCopy, null as any, blockMocks.testDatasetTree)).resolves.not.toThrow();
     });
 
     it("Checking failed copy of sequential datasets", async () => {
@@ -2222,7 +2222,7 @@ describe("Dataset Actions Unit Tests - Function copyDataSets", () => {
         const blockMocks = createBlockMocks();
         vscode.env.clipboard.writeText("");
         const errSpy = jest.spyOn(Gui, "errorMessage");
-        await expect(dsActions.pasteDataSetMembers(blockMocks.testDatasetTree, blockMocks.datasetSessionNode)).toEqual(Promise.resolve());
+        await expect(dsActions.pasteDataSetMembers(blockMocks.testDatasetTree, blockMocks.datasetSessionNode)).resolves.not.toThrow();
         expect(errSpy).toHaveBeenCalled();
     });
     it("Testing pasteDataSetMembers() fails and gives error message with empty clipboard", async () => {
@@ -2237,7 +2237,7 @@ describe("Dataset Actions Unit Tests - Function copyDataSets", () => {
             profile: blockMocks.imperativeProfile,
         });
         const errSpy = jest.spyOn(Gui, "errorMessage");
-        await expect(dsActions.pasteDataSetMembers(blockMocks.testDatasetTree, node)).toEqual(Promise.resolve());
+        await expect(dsActions.pasteDataSetMembers(blockMocks.testDatasetTree, node)).resolves.not.toThrow();
         expect(errSpy).toHaveBeenCalled();
     });
     it("Testing pasteDataSetMembers() succesfully runs pasteMember()", async () => {
@@ -2247,7 +2247,7 @@ describe("Dataset Actions Unit Tests - Function copyDataSets", () => {
         vscode.env.clipboard.writeText(JSON.stringify(getNodeLabels(blockMocks.pdsMemberNode)));
         const errSpy = jest.spyOn(dsActions, "pasteMember");
         errSpy.mockResolvedValueOnce(null);
-        await expect(dsActions.pasteDataSetMembers(blockMocks.testDatasetTree, blockMocks.pdsMemberNode)).toEqual(Promise.resolve());
+        await expect(dsActions.pasteDataSetMembers(blockMocks.testDatasetTree, blockMocks.pdsMemberNode)).resolves.not.toThrow();
     });
 
     it("Testing pasteDataSetMembers() successfully runs with multiple members", async () => {