+++ packages/zowe-explorer-api/CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to the "zowe-explorer-api" extension will be documented in t
 
 ### Bug fixes
 
+- Fixed issue where `zosmf` profiles did not respect the `protocol` property. [#2703](https://github.com/zowe/vscode-extension-for-zowe/issues/2703)
+
 ## `3.0.0-next.202402142205`
 
 ### New features and enhancements